refactor(cost): extract reference data adapter
