+++ src/services/portfolio_transaction_processing_service/app/infrastructure/cost_basis/__init__.py
@@ -1,5 +1,6 @@
-"""Expose typed infrastructure records for cost-basis effect staging."""
+"""Expose organized infrastructure adapters and records for cost-basis processing."""
 
+from .reference_data_repository import SqlAlchemyCostBasisReferenceDataRepository
 from .staged_effects import StagedCostEffects
 
-__all__ = ["StagedCostEffects"]
+__all__ = ["SqlAlchemyCostBasisReferenceDataRepository", "StagedCostEffects"]
