+++ src/services/portfolio_transaction_processing_service/app/infrastructure/sqlalchemy_unit_of_work.py
@@ -26,6 +26,7 @@ from .cashflow_processing_adapter import (
     CashflowStagingWorkflow,
 )
 from .cashflow_repository import SqlAlchemyCashflowRepository
+from .cost_basis import SqlAlchemyCostBasisReferenceDataRepository
 from .cost_processing_adapter import CostEffectsStager, CostProcessingCompatibilityAdapter
 from .cost_repository import CostCalculatorRepository
 from .pipeline_stage_processing_adapter import PipelineStageProcessingAdapter
@@ -145,6 +146,7 @@ class SqlAlchemyTransactionProcessingUnitOfWork:
         self._cost = CostProcessingCompatibilityAdapter(
             workflow=self._cost_workflow,
             repository=CostCalculatorRepository(session),
+            reference_data=SqlAlchemyCostBasisReferenceDataRepository(session),
             outbox_repository=outbox_repository,
         )
         self._cashflow = CashflowProcessingCompatibilityAdapter(
